Rework Client / Worker Remvoe WorkerMP and ClientMP

+++ src/scheduler/SchedulerWorkerMP.cpp
@@ -22,7 +22,7 @@ SchedulerWorkerMP::SchedulerWorkerMP(const std::function<std::shared_ptr<Worker>
 void SchedulerWorkerMP::processHardwareDetailMessage(const HardwareDetailMessage &msg) {
     //TODO No space in HardwareDetailMessage
     //TODO No OS Identifier in Spec
-    worker->setSpec({-1, msg.GetRamSize(), msg.GetCoreCount()});
+    worker->setSpec({(uint64_t)-1, msg.GetRamSize(), msg.GetCoreCount()});
 }
 
 void SchedulerWorkerMP::processPublicKeyAuthMessage(const PublicKeyAuthMessage &msg) {
